changed .js to .jsx, and organised components in it's own  directory
