+++ src/index.jsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom/client';
-import App from './App.js';
+import App from './App.jsx';
 import './global.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
